+++ cmd/kube-aggregator/pkg/apiserver/BUILD
@@ -61,7 +61,6 @@ go_library(
         "//pkg/genericapiserver/registry/generic/rest:go_default_library",
         "//pkg/genericapiserver/registry/rest:go_default_library",
         "//pkg/genericapiserver/server:go_default_library",
-        "//pkg/genericapiserver/server/filters:go_default_library",
         "//pkg/util/httpstream/spdy:go_default_library",
         "//pkg/util/workqueue:go_default_library",
         "//pkg/version:go_default_library",
@@ -73,6 +72,7 @@ go_library(
         "//vendor:k8s.io/apimachinery/pkg/util/runtime",
         "//vendor:k8s.io/apimachinery/pkg/util/wait",
         "//vendor:k8s.io/apiserver/pkg/endpoints/request",
+        "//vendor:k8s.io/apiserver/pkg/server/filters",
         "//vendor:k8s.io/client-go/rest",
         "//vendor:k8s.io/client-go/tools/cache",
         "//vendor:k8s.io/client-go/transport",
