+++ staging/src/k8s.io/apiserver/pkg/server/filters/doc.go
@@ -16,4 +16,4 @@ limitations under the License.
 
 // Package filters contains all the http handler chain filters which
 // are not api related.
-package filters // import "k8s.io/kubernetes/pkg/genericapiserver/server/filters"
+package filters // import "k8s.io/apiserver/pkg/server/filters"
